Refactor website redirect chart to support multiple entries
This is a breaking change for the website-redirect chart that allows for multiple
entries in one install of the chart.
redirects accepts multiple entries in an array that configure the server
to serve multiple virtual hosts with different redirects.
This as well sets up the ingress (with TLS) for all entered hostnames.

Signed-off-by: Maartje Eyskens <maartje@eyskens.me>